Submitted Project 02

+++ src/api/playlists/handler.js
@@ -1,9 +1,17 @@
+const AuthorizationError = require('../../exceptions/AuthorizationError');
 const BadRequestError = require('../../exceptions/BadRequestError');
 
 class PlaylistHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
+
+    this.postPlaylistsHandler = this.postPlaylistsHandler.bind(this);
+    this.getPlaylistsHandler = this.getPlaylistsHandler.bind(this);
+    this.deletePlaylistsHandler  = this.deletePlaylistsHandler.bind(this);
+    this.postPlaylistSongsHandler = this.postPlaylistSongsHandler.bind(this);
+    this.getPlaylistSongsHandler = this.getPlaylistSongsHandler.bind(this);
+    this.deletePlaylistSongsHandler = this.deletePlaylistSongsHandler.bind(this);
   }
   
   async postPlaylistsHandler(request, h) {
@@ -11,8 +19,9 @@ class PlaylistHandler {
       this._validator.validatePlaylistPayload(request.payload);
       
       const { name } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
       
-      const playlistId = await this._service.addPlaylist({ name });
+      const playlistId = await this._service.addPlaylist({ name, owner: credentialId });
       
       const response = h.response({
         status: 'success',
@@ -33,8 +42,7 @@ class PlaylistHandler {
         response.code(error.statusCode);
         return response;
       }
-      
-      // TODO: Mungkinkah?
+
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
@@ -45,25 +53,38 @@ class PlaylistHandler {
     }
   }
   
-  async getPlaylistsHandler() {
-    const playlists = await this._service.getPlaylists();
-    return {
-      status: 'success',
-      data: {
-        playlists,
-      },
-    };
+  async getPlaylistsHandler(request) {
+
+    const { id: credentialId } = request.auth.credentials;
+    const playlists = await this._service.getPlaylists(credentialId);
+
+    return playlists;
   }
   
   async deletePlaylistsHandler(request, h) {
     try {
-      const { id } = request.params;
-      await this._service.deletePlaylistById(id);
+      const { playlistId } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      console.log(playlistId, credentialId);
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+
+      await this._service.deletePlaylistById(playlistId);
+
       return {
         status: 'success',
         message: 'Palylist berhasil dihapus',
       };
+
     } catch (error) {
+      if(error instanceof AuthorizationError){
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
       if (error instanceof BadRequestError) {
         const response = h.response({
           status: 'fail',
@@ -84,27 +105,34 @@ class PlaylistHandler {
     }
   }
   
-  async postPlaylistsSongsHandler(request, h){
+  async postPlaylistSongsHandler(request, h){
     try{
       
       this._validator.validatePlaylistSongPayload(request.payload);
-      
+
+      const { id: credentialId } = request.auth.credentials;
       const { songId } = request.payload;
-      const {playlistId } = request.params;
+      const { playlistId } = request.params;
       
-      const playlistSongId = await this._service.addSongToPlaylist(playlistId, songId);
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+      const  playlistSongId = await this._service.addSongToPlaylist(playlistId, songId);
       
       const response = h.response({
         status: 'success',
         message: 'Lagu berhasil ditambahkan ke dalam playlist',
-        data: {
-          playlistSongId
-        }
       });
       response.code(201);
       return response;
       
     }catch(error) {
+      if(error instanceof AuthorizationError){
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        });
+        response.code(error.statusCode);
+        return response;
+      }
       
       if(error instanceof BadRequestError){
         const response = h.response({
@@ -113,7 +141,7 @@ class PlaylistHandler {
         });
         response.code(error.statusCode);
         return response;
-      }
+      } 
       
       const response = h.response({
         status: 'error',
@@ -125,17 +153,26 @@ class PlaylistHandler {
     }    
   }
   
-  async getPlaylistSongByIdHandler(){
+  async getPlaylistSongsHandler(request, h){
     try{
       const { playlistId } = request.params;
-      
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
       const songs = await this._service.getPlaylistSongById(playlistId);
       
-      return {
-        status: 'success',
-        data: songs
-      }
+      return songs;
+
     }catch(error){
+      if(error instanceof AuthorizationError){
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
       if (error instanceof BadRequestError) {
         const response = h.response({
           status: 'fail',
@@ -144,6 +181,8 @@ class PlaylistHandler {
         response.code(error.statusCode);
         return response;
       }
+
+      
       
       const response = h.response({
         status: 'error',
@@ -155,14 +194,31 @@ class PlaylistHandler {
     }
   }
 
-  async deletePlaylistSongByIdHandler(){
+  async deletePlaylistSongsHandler(request, h){
     try{
       const { playlistId } = request.params;
-      const {songId} = request.payload;
-      
+      const { songId } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyPlaylistOwner(playlistId, credentialId);
+
       await this._service.deleteSongFromPlaylist(playlistId, songId);
 
+      return {
+        status: 'success',
+        message: 'Lagu berhasil dihapus dari playlist'
+      }
+
     }catch (error) {
+      if(error instanceof AuthorizationError){
+        const response = h.response({
+          status: 'fail',
+          message: error.message
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
       if (error instanceof BadRequestError) {
         const response = h.response({
           status: 'fail',
@@ -171,6 +227,7 @@ class PlaylistHandler {
         response.code(error.statusCode);
         return response;
       }
+
       // Server ERROR!
       const response = h.response({
         status: 'error',
